pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Sergey Karatach <karatach1998@yandex.ru>"]

[tool.poetry.dependencies]
python = "3.7.10"
matplotlib = "^3.3"
sympy = "^1.7"
wolframalpha = "^4.1"
antlr4-python3-runtime = "^4.9"
yadisk = "^1.2.14"
PyYAML = "^5.4.1"
google-auth = "^1.27.1"
google-auth-oauthlib = "^0.4.3"
google-api-python-client = "^2.0.2"
google-auth-httplib2 = "^0.1.0"
pylint = "2.7.1"
python-telegram-bot = "^13.3"
PyPDF2 = "^1.26.0"
lxml = "^4.6.2"
mashumaro = "^2.0.2"
gunicorn = "^20.1.0"
Flask = "^1.1.2"
celery = "^5.0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pre-commit = "^2.10.1"
pylint = "2.7.1"
yapf = "^0.30.0"
toml = "^0.10.2"
pytest-cov = "^2.11.1"
mypy = "^0.812"

[mypy]
ignore_missing_imports = true
warn_no_return = false

[tool.pylint.message_control]
disable = ["R,C"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
